+++ hash/hash_crc32.c
@@ -38,10 +38,10 @@ uint32_t crc32(const char* s)
  */
 void test_crc32()
 {
-    assert(crc32("Hello World") == 1243066710);
-    assert(crc32("Hello World!") == 472456355);
-    assert(crc32("Hello world") == 2346098258);
-    assert(crc32("Hello world!") == 461707669);
+    assert(crc32("Hello World") == (const uint32_t)1243066710);
+    assert(crc32("Hello World!") == (const uint32_t)472456355);
+    assert(crc32("Hello world") == (const uint32_t)2346098258);
+    assert(crc32("Hello world!") == (const uint32_t)461707669);
     // printf("%" PRIu32 "\n", crc32("Hello World"));
     // printf("%" PRIu32 "\n", crc32("Hello World!"));
     // printf("%" PRIu32 "\n", crc32("Hello world"));
